backend/requirements.txt
fastapi
uvicorn
pymongo
pandas
numpy
scikit-learn
python-dotenv
google-generativeai
python-jose[cryptography]
bcrypt
python-multipart
email-validator
groq
